Rename fake_timers.js to fake-timers.js
This project uses kebab-case for .js filenames
