+++ lib/sinon.js
@@ -4,7 +4,7 @@ var behavior = require("./sinon/behavior");
 var createSandbox = require("./sinon/create-sandbox");
 var deprecated = require("./sinon/util/core/deprecated");
 var extend = require("./sinon/util/core/extend");
-var fakeTimers = require("./sinon/util/fake_timers");
+var fakeTimers = require("./sinon/util/fake-timers");
 var format = require("./sinon/util/core/format");
 var nise = require("nise");
 var Sandbox = require("./sinon/sandbox");
