+++ lib/sinon/sandbox.js
@@ -8,7 +8,7 @@ var isPropertyConfigurable = require("./util/core/is-property-configurable");
 var isNonExistentOwnProperty = require("./util/core/is-non-existent-own-property");
 var sinonMatch = require("./match");
 var sinonAssert = require("./assert");
-var sinonClock = require("./util/fake_timers");
+var sinonClock = require("./util/fake-timers");
 var sinonMock = require("./mock");
 var sinonSpy = require("./spy");
 var sinonStub = require("./stub");
