+++ lib/sinon/util/fake-timers.js
